Add earliest partition date stat to table stats collector (#205)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

This PR adds the earliest partition date on a table as a stat to the
table stats collector.

For tables where the retention policy has been set on a string type
column, the stat is retrieved based on the retention column name, since
this stat will be used as a comparison point for Openhouse alerting to
identify if older partitions should have been cleaned up through
retention jobs.
For tables that are partitioned on timestamp, it searches for the
partition column name that is `Timestamp`. Openhouse only allows a table
to be partitioned on 1 timestamp type column so we can identify that as
the column set for retention.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [x] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [x] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

Added unit testing.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.