Feature-Toggle Use case: Allow preserved tblprops to be modified on Table basis (#189)
This PR demonstrates how annotation-based feature-toggle can help with
allowing certain "preserved table properties" to be modifiable by
privileged users of certain tables. Common use cases for this is when
there are high-impact features that team is not feeling confident to
rollout to all tables, we can use feature-toggle to allow some tables to
be able to use this first.

Here for demonstration purpose, we show how `openhouse.tableType` can
become writable if the feature `enable_tabletype` is toggled on for
certain tables, in a e2e unit test setup.

This PR also contains some refactoring required to make this work.
Specifically, for the interface used to manage preserved keys, we need
to introduce table-awareness of it so that it can make decision based on
the table identifier. There are also a few test cases adjustment (e.g.
moved from e2e-type into mock as I see fit).

Every time when developers need to add new table properties into this
toggle-managed process, one needs to do the following:

- Use `TblPropsToggleRegistry` to register feature and its corresponding
table properties. For merge-on-read feature for example, one will need
to add the 3 keys (`write.delete.mode`, `write.update.mode`,
`write.merge.mode` ) into the map and with the value as pre-defined
featureId, say `enable_mor`. This is intentionally designed to be a code
change.
- Deploy the code change above.
- Update feature toggle's backend to allow some table with specific
naming patterns to have `enable_mor` activated