Relax LocalStorageClient's LocalFileSystem check (#215)
## Summary

In `LocalStorageClient` we have a check to enforce that the `FileSystem`
instance which is created is a `LocalFileSystem`. In some cases, for
example when running within the Trino context (e.g. for an integration
test), the `FileSystem` instance may be a _wrapper_ around a
`LocalFileSystem`. This should still be considered valid.

For reference, here is the code within Trino which will force that all
`FileSystem` instances get wrapper inside of a custom wrapper
`FileSystemWrapper`, including `LocalFileSystem`:

https://github.com/trinodb/trino/blob/f1a8feb8e4d38fc771c0d09f85c93dae3886456a/lib/trino-hdfs/src/main/java/io/trino/hdfs/TrinoFileSystemCache.java#L158-L160

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [X] Tests

## Testing Done
- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [X] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

New unit test in `LocalStorageClientTest`

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

(none apply)