Updating  tables-test-fixture uber jar to not have any unshaded dependencies (#233)
## Summary

The tables-test-fixture uber jar brought in a lot of unrelocated
dependencies some of which are unnecessary but caused classpath
conflicts while leveraging it for testing. So the objective of this PR
is to
1. exclude all dependencies that are not required.
2. relocate the ones that are still needed to avoid classpath conflicts.

The exclusions and relocations were identified through a combination of
trial and error and through dependencyInsights.
springframework is still pending relocation and is kept out of scope for
this PR. This will be addressed if there is a need for the same. This
specific relocation caused autowired beans to be not recognized despite
merging or appending the transformed properties or by modifying
spring.factories file. Added a TODO to address it if it comes up as an
issue.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [X] Tests

Testing by depending on this jar on another product and ensuring the
tests go through.

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

Jars that are present post exclusion and relocation
![Screenshot 2024-10-17 at 5 57
33 PM](https://github.com/user-attachments/assets/c0f05910-3ecb-4153-bb88-8934ba09ea88)


# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.