Remove some relocations and exclusions in tables-test-fixture (#244)
## Summary

This PR adds back some exclusions and undoes some relocations that were
done as part of PR #233.
The uber jar contains some additional dependencies but minimum ones that
are required to run tables-test-fixtures test.


<img width="337" alt="Screenshot 2024-10-30 at 1 53 01 PM"
src="https://github.com/user-attachments/assets/d599bbb6-7a38-46e6-bfc2-2d2f627852f0">


Reasons for retaining the jars:
1. Relocation or exclusion of hibernate, ehcache, h2 makes dependencies
to be unavailable for hts as part of the tests.
2. Relocation or exclusion of zaxxer makes wrong primary bean to be
picked up for hts as part of the tests.
3. javaxx relocation or exclusion causes class not found errors as
configs are required to be overridden in tomcat which inturn cant be
relocated without springframework relocation. Same applies for springdoc
relocation and io.micrometer as well

Keeping the above out of scope for this PR and can be explored further
to identify the right set of exclusions or relocations that need to
happen together.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [X] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [X] Some other form of testing like staging or soak time in
production. Please explain.
Tested with snapshot jar and a new module that Sushant provided.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.