Convert interval parameter to cron schedule for self serve replication (#238)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

This PR adds cron schedule to the replication config for the self serve
replication API and converts the interval parameter to a cron schedule.
Cron schedule is needed as an input for the replication job that
performs the cross cluster data copy. The library `cronutils` is added
to perform the cron expression generation.

The interval parameter can be validated as `12H`, `1D`, `2D`, `3D` and
the cron schedule is generated based on:
- 12H schedule dictates replication should trigger every 12 hours
- The cron schedule should have an X hour to start from midnight where X
can range from 0-23 and Y minute where Y is in [0, 15, 30, 45]. E.g
12hours can lead to schedule “0 30 3/12 ? * * *” -> trigger at 3:30am
and 3:30pm every day.
- If interval is not provided, a daily replication schedule should be
set up with daily schedules with X hour starting from midnight.
- X is randomized from 0-23 and Y is randomized as a value of [0, 15,
30, 45] to spread out the cron and avoid job clusters around a time.

Examples of:
12H: `0 15 23/12 ? * * *`
1D: `0 45 10 ? * * *`
2D: `0 30 8 ? * */2 *`
3D: `0 0 8 ? * */3 *`

## Changes

- [x] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [x] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [x] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [x] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

Unit tests added to check that the cron schedule was created
successfully.

POST http://localhost:8000/v1/databases/u_tableowner/tables for a 1D
schedule:
```
{
    "policies": {
        "sharingEnabled": "true",
        "replication": {
            "config": [
                {
                    "destination": "clusterA"                
                 }
            ]
        }
    }
}
```
response:
```
        "replication": {
            "config": [
                {
                    "destination": "clusterA",
                    "interval": "1D",
                    "cronSchedule": "0 0 10 ? * * *"
                }
            ]
        }
```
12H schedule:
```
"replication": {
            "config": [
                {
                    "destination": "clusterA",
                    "interval": "12H",
                    "cronSchedule": "0 0 23/12 ? * * *"
                }
            ]
        }
```
# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.