Enlarge the cache used to avoid retry from internal-catalog towards HTS (#250)
## Summary

This change is identified as one of the action item from investigation
of a table corruption issue caused by retry towards House table service.

What we wanted to achieve here is, within a lifespan of a client-started
transaction on the server side(after taking all the retry possibility
into the consideration), all of the retry attempts will be intentionally
blocked from the internal catalog's perspective.

By running the audit log internally, we have seen ~50K put requests on
the table-service and that puts ~170 PUT request for every 5 minute
time-window. Counting in the future growth, we reach 1k as the cache
size.