[PR1] refactoring :apps build and adding spark3.5/iceberg1.5 artifact (#255)
## Summary
This is Part 1 of 2 for enabling compaction based MoR.

1. bring spark3.5 for creating delete files in testing
2. enable delete file compaction in spark3.1/spark3.5 and add unittests
for such in spark3.5

problem: enabling Merge-on-Read for customer tables requires the ability
to compact delete files. Compacting delete files is supported in
spark3.1/iceberg1.2 but requires us to test it. We cannot test it,
because creating delete files is not supported in spark3.1 (only
compacting existing ones).

solution: we add spark3.5/iceberg1.5 compaction application, which
extends the existing compaction test suite. in a follow up PR, we will
add tests which create delete files and compacts the delete files under
the latest versions.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [X] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [X] Refactoring
- [ ] Documentation
- [X] Tests

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [X] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

all tests running on spark3.1 and spark3.5:

![image](https://github.com/user-attachments/assets/65330bbb-b440-46b1-8fe9-92162d5e7a94)

correct spark version resolutions for test runtime:

![image](https://github.com/user-attachments/assets/b0669745-0e0e-4bf3-a02b-c500c1053fb6)

correct spark version resolutions for runtime:

![image](https://github.com/user-attachments/assets/48725130-14f8-4fc9-8da3-05f8480f657c)

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.