Fix Gson non escaped chars in Table Policy (#264)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

[Issue](https://github.com/linkedin/openhouse/issues/#nnn)] 
Currently the display of the policies are in the following format
```
policies                                  
{
  "retention": {
    "count": 180,
    "granularity": "DAY",
    "columnPattern": {
      "columnName": "datepartition",
      "pattern": "\u0027yyyy-MM-dd\u0027"
    }
  },
  "sharingEnabled": true
}
```
we would like to escape the `\u0027` unicode to the single quotation
marks to avoid confusion from the users

This is because during the write pathGson natively escapes the dangerous
json characters during writing

https://github.com/google/gson/blob/main/gson/src/main/java/com/google/gson/stream/JsonWriter.java#L199

During the read path returned from iceberg, all UTF-8 encodings will no
deserialize this unicode

Using `disableHtmlEscaping` to save the string without the unicode

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [x] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [x] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [x] Added new tests for the changes made.
- [x] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.