[PR2] Bin-pack delete files flag, MoR feature validation (only spark3.5), and observability (#265)
## Summary
To support row level updates prevalent in CDC use cases, we will enable
MoR in OpenHouse. Enabling MoR has a risk of delete file accumulation
which affects storage quota utilization and compute performance.
Therefore we must mitigate this risk with efficient compaction of delete
files.

Here we are implementing:
- assertions that delete file compaction is enabled in the OpenHouse
stack, indirectly with spark3.5 due to spark3.1 inability to create
delete files
- asserting feature gaps which we have researched about delete file
compaction
- key observability metrics for this delete file compaction
- exposing the flag to trigger this compaction

minimal tests to be added:
```java
testDeleteFilesCanBeCreated();
testCompactionCanRemovePositionDeleteFiles();
testCompactionCanRemoveEqualityDeleteFiles();
```

tests which I will add in a PR3. these aim to reinforce our foundational
knowledge or cover known edge cases:
```java
// this validates our understanding of datafile sequence number being the primary mechanism to remove delete files as is documented here: https://docs.google.com/document/d/11d-cIUR_89kRsMmWnEoxXGZCvp7L4TUmPJqUC60zB5M/edit?tab=t.0#bookmark=id.a485992lwbxz 
testSnapshotExpirationCanRemovePositionDeleteFiles(); 
testSnapshotExpirationCanRemoveEqualityDeleteFiles(); 

// these are documented issues where delete files may accumulate, needs understanding of data sequence number
testDeleteFilesNotExpirableBecauseOldDatafile();
testDeleteFilesNotExpirableBecauseDangling();

// this is our recovery process to implement in the future  via spark application in case delete files do accumulate from above
testProcessToCompactNonExpirableDeleteFiles();
```

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

[Issue](https://github.com/linkedin/openhouse/issues/#nnn)] Briefly
discuss the summary of the changes made in this
pull request in 2-3 lines.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [x] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [x] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [x] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.